Refactor: Move source management renderlayer code into generic sliceview class.
The backend renderlayer classes in point (vectorgraphics) and volume simply override the sources member and use the default SliceView RenderLayer constructor. The frontend classes work similarly. In addition, some volume specific code was removed from Volume/RenderLayer and placed in SliceView/RenderLayer
