+++ src/neuroglancer/sliceview/renderlayer.ts
@@ -28,96 +28,6 @@ import {makeWatchableShaderError, WatchableShaderError} from 'neuroglancer/webgl
 import {getShaderType} from 'neuroglancer/webgl/shader_lib';
 import {RpcId, SharedObject} from 'neuroglancer/worker_rpc';
 
-
-export class SliceViewShaderBuffers extends RefCounted {
-  outputVertexIndices: Buffer;
-  vertexBasePositions: number[];
-  vertexIndices: Int32Array;
-  constructor(gl: GL) {
-    super();
-    this.outputVertexIndices = this.registerDisposer(
-        Buffer.fromData(gl, new Float32Array([0, 1, 2, 3, 4, 5]), gl.ARRAY_BUFFER, gl.STATIC_DRAW));
-
-    // This specifies the original, "uncorrected" vertex positions.
-    // var vertexBasePositions = [
-    //   0, 0, 0,
-    //   1, 0, 0,
-    //   0, 1, 0,
-    //   0, 0, 1,
-    //   1, 0, 1,
-    //   1, 1, 0,
-    //   0, 1, 1,
-    //   1, 1, 1,
-    // ];
-
-    // This specifies the "corrected" vertex positions.
-    this.vertexBasePositions = [
-      0, 0, 0,  //
-      1, 0, 0,  //
-      0, 1, 0,  //
-      1, 1, 0,  //
-      0, 0, 1,  //
-      1, 0, 1,  //
-      0, 1, 1,  //
-      1, 1, 1,  //
-    ];
-
-    // correct_index, vertex_position, uncorrected_index
-    // 0:  0, 0, 0   0
-    // 1:  1, 0, 0   1
-    // 2:  0, 1, 0   2
-    // 4:  0, 0, 1   3
-    // 5:  1, 0, 1   4
-    // 3:  1, 1, 0   5
-    // 6:  0, 1, 1   6
-    // 7:  1, 1, 1   7
-
-    // This maps uncorrected vertex indices to corrected vertex indices.
-    let vertexUncorrectedToCorrected = [0, 1, 2, 4, 5, 3, 6, 7];
-
-    // This maps corrected vertex indices to uncorrected vertex indices.
-    let vertexCorrectedToUncorrected = [0, 1, 2, 5, 3, 4, 6, 7];
-
-
-    // Page 666
-    let vertexBaseIndices = [
-      0, 1, 1, 4, 4, 7, 4, 7,  //
-      1, 5, 0, 1, 1, 4, 4, 7,  //
-      0, 2, 2, 5, 5, 7, 5, 7,  //
-      2, 6, 0, 2, 2, 5, 5, 7,  //
-      0, 3, 3, 6, 6, 7, 6, 7,  //
-      3, 4, 0, 3, 3, 6, 6, 7,  //
-    ];
-
-    // Determined by looking at the figure and determining the corresponding
-    // vertex order for each possible front vertex.
-    let vertexPermutation = [
-      0, 1, 2, 3, 4, 5, 6, 7,  //
-      1, 4, 5, 0, 3, 7, 2, 6,  //
-      2, 6, 0, 5, 7, 3, 1, 4,  //
-      3, 0, 6, 4, 1, 2, 7, 5,  //
-      4, 3, 7, 1, 0, 6, 5, 2,  //
-      5, 2, 1, 7, 6, 0, 4, 3,  //
-      6, 7, 3, 2, 5, 4, 0, 1,  //
-      7, 5, 4, 6, 2, 1, 3, 0,  //
-    ];
-
-    let vertexIndices: number[] = [];
-    for (var p = 0; p < 8; ++p) {
-      for (var i = 0; i < vertexBaseIndices.length; ++i) {
-        const vertexPermutationIndex = vertexCorrectedToUncorrected[p] * 8 + vertexBaseIndices[i];
-        vertexIndices.push(vertexUncorrectedToCorrected[vertexPermutation[vertexPermutationIndex]]);
-      }
-    }
-
-    this.vertexIndices = new Int32Array(vertexIndices);
-  }
-
-  static get(gl: GL) {
-    return gl.memoize.get('SliceViewShaderBuffers', () => new SliceViewShaderBuffers(gl));
-  }
-};
-
 const tempVec3 = vec3.create();
 const tempVec3b = vec3.create();
 const tempMat4 = mat4.create();
@@ -125,11 +35,12 @@ const tempMat4 = mat4.create();
 export abstract class RenderLayer extends GenericRenderLayer {
   chunkManager: ChunkManager;
   sources: SliceViewChunkSource[][]|null = null;
+  sourceIds: number[][] = [];
   shader: ShaderProgram|undefined = undefined;
   shaderUpdated = true;
   rpcId: RpcId|null = null;
   shaderError: WatchableShaderError;
-  constructor(chunkManager: ChunkManager, spec: SliceViewChunkSpecification, {
+  constructor(chunkManager: ChunkManager, sources: SliceViewChunkSource[][], {
     shaderError = makeWatchableShaderError(),
   } = {}) {
     super();
@@ -137,8 +48,20 @@ export abstract class RenderLayer extends GenericRenderLayer {
     this.shaderError = shaderError;
     shaderError.value = undefined;
     this.chunkManager = chunkManager;
+    console.log(sources);
+    this.sources = sources;
+    console.log(this.sources);
     let gl = this.gl;
 
+    for (let alternatives of sources) {
+      let alternativeIds: number[] = [];
+      this.sourceIds.push(alternativeIds);
+      for (let source of alternatives) {
+        alternativeIds.push(source.rpcId!);
+      }
+    }
+
+    let spec = this.sources[0][0].spec;
     let {chunkLayout} = spec;
 
     const voxelSize = this.voxelSize =
@@ -196,7 +119,7 @@ export abstract class RenderLayer extends GenericRenderLayer {
   getShaderKey() { return ''; }
 
   getShader() {
-    let key = this.getShaderKey(); 
+    let key = this.getShaderKey();
     return this.gl.memoize.get(key, () => this.buildShader());
   }
 
@@ -208,9 +131,9 @@ export abstract class RenderLayer extends GenericRenderLayer {
 
   abstract defineShader(builder: ShaderBuilder): void
 
-  abstract beginSlice(_sliceView: SliceView): ShaderProgram 
+  abstract beginSlice(_sliceView: SliceView): ShaderProgram
 
-  abstract endSlice(shader: ShaderProgram): void 
+  abstract endSlice(shader: ShaderProgram): void
 
-  abstract draw(sliceView: SliceView): void 
+  abstract draw(sliceView: SliceView): void
 }
