+++ src/neuroglancer/sliceview/volume/backend.ts
@@ -94,22 +94,7 @@ export abstract class VolumeChunkSource extends SliceViewChunkSource implements
 
 @registerSharedObject(VOLUME_RENDERLAYER_RPC_ID)
 export class RenderLayer extends SliceViewRenderLayer implements RenderLayerInterface {
-  rpcId: number;
   sources: VolumeChunkSource[][];
-
-  constructor(rpc: RPC, options: any) {
-    super(rpc, options);
-    let sources = this.sources = new Array<VolumeChunkSource[]>();
-    for (let alternativeIds of options['sources']) {
-      let alternatives = new Array<VolumeChunkSource>();
-      sources.push(alternatives);
-      for (let sourceId of alternativeIds) {
-        let source: VolumeChunkSource = rpc.get(sourceId);
-        this.registerDisposer(source.addRef());
-        alternatives.push(source);
-      }
-    }
-  }
 }
 
 /**
